+++ config/routes.rb
@@ -1,11 +1,12 @@
 # frozen_string_literal: true
 
 DataPorter::Engine.routes.draw do
-  resources :imports, only: %i[index new create show destroy] do
+  resources :imports, path: "/", only: %i[index new create show destroy] do
     member do
       post :parse
       post :confirm
       post :cancel
+      post :back_to_mapping
       post :dry_run
       patch :update_mapping
       get :status
@@ -13,5 +14,5 @@ DataPorter::Engine.routes.draw do
     end
   end
 
-  resources :mapping_templates, except: :show
+  resources :mapping_templates
 end